agents/real-estate-extractor/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
xlrd==2.0.1

# AI and NLP
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
tiktoken==0.5.2

# Data validation and processing
jsonschema==4.20.0
python-dateutil==2.8.2

# File handling
aiofiles==23.2.1
Pillow==10.1.0

# Utilities
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2 